Issue 1903 update
Added more class errors from asgi to the error_module
